MyAddressBookAPI/MyAddressBookAPI/Controllers/ContactsController.cs
using System;
using System.Collections.Generic;
using System.Data;
using System.Data.Entity;
using System.Data.Entity.Infrastructure;
using System.Linq;
using System.Net;
using System.Net.Http;
using System.Threading.Tasks;
using System.Web.Http;
using System.Web.Http.Description;
using MyAddressBookAPI.Models;

namespace MyAddressBookAPI.Controllers
{
    /// <summary>
    /// Rest WebApi controller for MyAddressBook app.
    /// Originally code was scaffolded.  Altered to
    /// allow dependency injection.
    /// </summary>
    public class ContactsController : ApiController
    {
        private IDbContext db;

        public ContactsController()
        {
            db = new MyAddressBookDBContext();
        }

        public ContactsController(IDbContext dbInjected)
        {
            db = dbInjected;
        }

        // GET: api/Contacts
        public IQueryable<Contact> GetContacts()
        {
            return db.Query<Contact>();
        }

        // GET: api/Contacts/5
        [ResponseType(typeof(Contact))]
        public async Task<IHttpActionResult> GetContact(int id)
        {
            Contact contact = await db.FindAsync<Contact>(id);

            if (contact == null)
            {
                return NotFound();
            }

            return Ok(contact);
        }

        // PUT: api/Contacts/5
        [ResponseType(typeof(void))]
        public async Task<IHttpActionResult> PutContact(int id, Contact contact)
        {
            if (!ModelState.IsValid)
            {
                return BadRequest(ModelState);
            }

            if (id != contact.ContactId)
            {
                return BadRequest();
            }

            db.Update<Contact>(contact);

            try
            {
                await db.SaveChangesAsync();
            }
            catch (DbUpdateConcurrencyException)
            {
                if (!ContactExists(id))
                {
                    return NotFound();
                }
                else
                {
                    throw;
                }
            }

            return StatusCode(HttpStatusCode.NoContent);
        }

        // POST: api/Contacts
        [ResponseType(typeof(Contact))]
        public async Task<IHttpActionResult> PostContact(Contact contact)
        {
            if (!ModelState.IsValid)
            {
                return BadRequest(ModelState);
            }

            db.Add<Contact>(contact);
            await db.SaveChangesAsync();

            return CreatedAtRoute("DefaultApi", new { id = contact.ContactId }, contact);
        }

        // DELETE: api/Contacts/5
        [ResponseType(typeof(Contact))]
        public async Task<IHttpActionResult> DeleteContact(int id)
        {
            Contact contact = await db.FindAsync<Contact>(id);

            if (contact == null)
            {
                return NotFound();
            }

            db.Remove<Contact>(contact);
            await db.SaveChangesAsync();

            return Ok(contact);
        }

        protected override void Dispose(bool disposing)
        {
            if (disposing)
            {
                db.Dispose();
            }
            base.Dispose(disposing);
        }

        private bool ContactExists(int id)
        {
            return db.Query<Contact>().Count(e => e.ContactId == id) > 0;
        }
    }
}